refactor(events): move orchestration sse tests into aura-events
The orchestration SSE types already live in aura-events; their
wire-format and backward-compat tests sat in the aura crate's
orchestration::stream_events module, which only re-exported them.
Move the tests next to the types and delete the module, repointing
the re-exports straight at aura-events. No external callers used it.

Ref: #221
